+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Users can now edit the size of SAP log volumes for Power VS
 - When using the search bar in the hamburger menu, results will be returned based on documentation text
 
 ### Fixes